# src/middelware/tsoa/swagger.json
{
	"components": {
...
		"schemas": {
			"InternalServerError": {
				"properties": {
					"stack": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"VerifyResend": {
				"properties": {
					"preferred_username": {
						"type": "string"
					}
				},
				"required": [
					"preferred_username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PasswordResetRequest": {
				"properties": {
					"preferred_username": {
						"type": "string"
					}
				},
				"required": [
					"preferred_username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PasswordReset": {
				"properties": {
					"password": {
						"type": "string"
					}
				},
				"required": [
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SignUp": {
				"properties": {
					"preferred_username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"given_name": {
						"type": "string"
					},
					"family_name": {
						"type": "string"
					}
				},
				"required": [
					"preferred_username",
					"password",
					"given_name",
					"family_name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SignIn": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ClientPost": {
				"properties": {
					"client_name": {
						"type": "string"
					},
					"client_secret": {
						"type": "string"
					},
					"redirect_uri": {
						"type": "string"
					}
				},
				"required": [
					"client_name",
					"client_secret",
					"redirect_uri"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"oauth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://localhost:3000/oidc/auth",
						"scopes": {
							"openid": "modify things",
							"profile": "read things",
							"email": "Get user info"
						}
					}
				}
			},
			"openid": {
				"type": "openIdConnect",
				"openIdConnectUrl": "http://localhost:3000/oidc/.well-known/openid-configuration"
			},
			"google": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://localhost:7000/auth/google",
						"scopes": {}
					}
				}
			},
			"facebook": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://localhost:7000/auth/facebook",
						"scopes": {}
					}
				}
			}
...
	},
	"info": {
		"title": "micro-identity",
		"version": "0.0.1",
		"description": "Boilerplate for creating identity microservice",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Martin Greyling"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/account/verify_resend": {
			"post": {
				"operationId": "VerifyResend",
				"responses": {
					"200": {
						"description": "Verification Message Resent!",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"Verification Message API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Account"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyResend"
							}
						}
					}
				}
			}
		},
		"/account/verify": {
			"get": {
				"operationId": "VerifyGet",
				"responses": {
					"200": {
						"description": "Account Verified!",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"Verify API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Account"
				],
				"security": [
					{
						"jwt-query": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/account/password_reset_request": {
			"post": {
				"operationId": "PasswordRequestPost",
				"responses": {
					"200": {
						"description": "Account Verified!",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"Verify API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Account"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PasswordResetRequest"
							}
						}
					}
				}
			}
		},
		"/account/password_reset": {
			"post": {
				"operationId": "PasswordResetPost",
				"responses": {
					"200": {
						"description": "Account Verified!",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"Verify API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Account"
				],
				"security": [
					{
						"jwt-query": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PasswordReset"
							}
						}
					}
				}
			}
		},
...
		"/auth/signup": {
			"post": {
				"operationId": "SignUpPost",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"Signup API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignUp"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"get": {
				"operationId": "LoginGet",
				"responses": {
					"200": {
						"description": "User Logged In!",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"Signin API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "LoginPost",
				"responses": {
					"200": {
						"description": "User Logged In!",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"Signin API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignIn"
							}
						}
					}
				}
			}
		},
		"/auth/logout": {
			"get": {
				"operationId": "LogoutGet",
				"responses": {
					"200": {
						"description": "User Logged In!",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"Signin API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "LogoutPost",
				"responses": {
					"200": {
						"description": "User Logged In!",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"Signin API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignIn"
							}
						}
					}
				}
			}
		},
...
		"/client/{clientName}": {
			"get": {
				"operationId": "GetClient",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"Oauth Client API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Clients"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "clientName",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteClient",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"Oauth Client API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Clients"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "clientName",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "softDelete",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					}
				]
			}
		},
		"/client": {
			"get": {
				"operationId": "GetClients",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"Oauth Client API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Clients"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateClient",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"Oauth Client API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Clients"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ClientPost"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateClient",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"Oauth Client API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Clients"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ClientPost"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}